skel: Added yaml get, append, and replace
New functionality to ease the modification of existing YAML files.
Can now retrieve, append, and replace configurations in YAML files.
This is useful for when we need to add new configurations to YAML files
for newly generated templates.

Added addYamlConf to add to or modify YAML
Allows for easy changes to YAML files to add new lines that are required
after templates are generated.